Add session_key attribute to util module
It is not required as a ipyEncodeAndSend argument any more, thus we do not have
to find a way to propagate it.

`session_key` is private and is set through `setSessionKey`.

No `getSessionKey` has been created yet as we have no use of it, but it can be
easily added.